reactor: move lldb related part to another mod
