+++ Cargo.toml
@@ -10,6 +10,7 @@ lldb = { git = "https://github.com/endoli/lldb.rs.git", version = "0.0.11" }
 serde = { version = "1.0.215", features = ["derive"] }
 serde_json = "1.0.133"
 tempfile = "3.14.0"
+thiserror = "2.0.3"
 uuid = { version = "1.11.0", features = ["v4"] }
 
 [workspace]
