+++ src/process_remote.rs
@@ -1,0 +1,11 @@
+mod lldb;
+
+#[derive(Debug, thiserror::Error)]
+pub enum ProcessRemoteError {
+    #[error("byte order mismatch")]
+    ByteOrderMismatch,
+    #[error("pointer size mismatch")]
+    PointerSizeMismatch,
+}
+
+pub use lldb::get_buffer;
